+++ fabric/message_bus/messages/add_update_reservation_record.py
@@ -24,28 +24,32 @@
 #
 # Author: Komal Thareja (kthare10@renci.org)
 """
-Implements Avro representation of a Get Reservations State Message
+Implements Avro representation of an UpdateAddReservation Message
 """
-from typing import List
 from uuid import uuid4
+
+from fabric.message_bus.messages.reservation_mng import ReservationMng
+from fabric.message_bus.messages.ticket_reservation_avro import TicketReservationAvro
 
 from fabric.message_bus.messages.auth_avro import AuthAvro
 from fabric.message_bus.messages.message import IMessageAvro
 
 
-class GetReservationsStateRequestAvro(IMessageAvro):
+class AddUpdateReservationRecord(IMessageAvro):
     """
-    Implements Avro representation of a Get Reservations State Message
+    Implements Avro representation of an UpdateAddReservation Message
     """
     # Use __slots__ to explicitly declare all data members.
-    __slots__ = ["name", "message_id", "guid", "auth", "reservation_ids", "callback_topic", "id_token", "id"]
+    __slots__ = ["name", "message_id", "guid", "auth", "reservation_obj", "reservation_id",
+                 "callback_topic", "id_token", "id"]
 
     def __init__(self):
-        self.name = IMessageAvro.GetReservationsStateRequest
+        self.name = None
         self.message_id = None
         self.guid = None
         self.auth = None
-        self.reservation_ids = None
+        self.reservation_obj = None
+        self.reservation_id = None
         self.callback_topic = None
         self.id_token = None
         # Unique id used to track produce request success/failures.
@@ -58,17 +62,26 @@ class GetReservationsStateRequestAvro(IMessageAvro):
         For this reason we must provide conversion from dict to our class for de-serialization
         :param value: incoming message dictionary
         """
-        if value['name'] != IMessageAvro.GetReservationsStateRequest:
-            raise Exception("Invalid message")
-        self.message_id = value['message_id']
-        self.guid = value['guid']
-        self.callback_topic = value['callback_topic']
-        self.reservation_ids = value.get("reservation_ids", None)
-        self.id_token = value.get("id_token", None)
+        self.message_id = value.get('message_id', None)
+        self.guid = value.get('guid', None)
+        self.callback_topic = value.get('callback_topic', None)
+        self.reservation_id = value.get('reservation_id', None)
+
+        if value.get("reservation_obj", None) is not None:
+            res_value = value.get("reservation_obj", None)
+            class_name = res_value.get('name', None)
+            if class_name == TicketReservationAvro.__name__:
+                self.reservation_obj = TicketReservationAvro()
+                self.reservation_obj.from_dict(res_value)
+            elif class_name == ReservationMng.__name__:
+                self.reservation_obj = ReservationMng()
+                self.reservation_obj.from_dict(res_value)
 
         if value.get('auth', None) is not None:
             self.auth = AuthAvro()
             self.auth.from_dict(value['auth'])
+
+        self.id_token = value.get('id_token')
 
     def to_dict(self) -> dict:
         """
@@ -78,20 +91,24 @@ class GetReservationsStateRequestAvro(IMessageAvro):
         """
         if not self.validate():
             raise Exception("Invalid arguments")
+
         result = {
             "name": self.name,
             "message_id": self.message_id,
             "guid": self.guid,
             "callback_topic": self.callback_topic
         }
-        if self.id_token is not None:
-            result["id_token"] = self.id_token
 
+        if self.reservation_id is not None:
+            result['reservation_id'] = self.reservation_id
         if self.auth is not None:
             result['auth'] = self.auth.to_dict()
 
-        if self.reservation_ids is not None:
-            result['reservation_ids'] = self.reservation_ids
+        if self.reservation_obj is not None:
+            result['reservation_obj'] = self.reservation_obj.to_dict()
+
+        if self.id_token is not None:
+            result['id_token'] = self.id_token
 
         return result
 
@@ -102,38 +119,54 @@ class GetReservationsStateRequestAvro(IMessageAvro):
         return self.message_id
 
     def get_message_name(self) -> str:
+        """
+        Returns the message name
+        """
         return self.name
 
     def get_callback_topic(self) -> str:
+        """
+        Returns the callback topic
+        """
         return self.callback_topic
 
     def get_id(self) -> str:
+        """
+        Returns the id
+        """
         return self.id.__str__()
 
-    def get_reservation_ids(self) -> List[str]:
+    def get_reservation(self) -> ReservationMng:
         """
-        Return list of reservation ids
+        Returns the reservation object
+        @return reservation
         """
-        return self.reservation_ids
+        return self.reservation_obj
+
+    def get_reservation_id(self) -> str:
+        """
+        Return reservation id
+        """
+        return self.reservation_id
 
     def get_id_token(self) -> str:
         """
-        Return identity token
+        Return id token
         """
         return self.id_token
 
     def __str__(self):
-        return "name: {} message_id: {} guid: {} auth: {} reservation_ids: {} callback_topic: {} id_token: {}".format(
-            self.name, self.message_id, self.guid, self.auth, self.reservation_ids, self.callback_topic, self.id_token)
+        return "name: {} message_id: {} guid: {} auth: {} reservation_obj: {} reservation_id: {} callback_topic: {}" \
+               " id_token: {}".format(self.name, self.message_id, self.guid, self.auth, self.reservation_obj,
+                                      self.reservation_id, self.callback_topic, self.id_token)
 
     def validate(self) -> bool:
         """
         Check if the object is valid and contains all mandatory fields
         :return True on success; False on failure
         """
-        ret_val = super().validate()
-
-        if self.guid is None or self.auth is None or self.callback_topic is None or self.reservation_ids is None:
+        ret_val = True
+        if not super().validate() or self.guid is None or self.auth is None or self.callback_topic is None or \
+                self.reservation_obj is None:
             ret_val = False
-
         return ret_val
