+++ fabric/message_bus/messages/extend_ticket_avro.py
@@ -27,11 +27,11 @@
 Implements Avro representation of an Extend Ticket Message
 """
 
-from fabric.message_bus.messages.message_base import MessageBaseAvro
+from fabric.message_bus.messages.reservation_or_delegation_record import ReservationOrDelegationRecord
 from fabric.message_bus.messages.message import IMessageAvro
 
 
-class ExtendTicketAvro(MessageBaseAvro):
+class ExtendTicketWithReservationOrDelegationRecord(ReservationOrDelegationRecord):
     """
     Implements Avro representation of an Extend Ticket Message
     """
