+++ fabric/message_bus/messages/get_pool_info_avro.py
@@ -26,99 +26,14 @@
 """
 Implements Avro representation of a Get Pool Info Message
 """
-from uuid import uuid4
-
-from fabric.message_bus.messages.auth_avro import AuthAvro
+from fabric.message_bus.messages.request_by_id_record import RequestByIdRecord
 from fabric.message_bus.messages.message import IMessageAvro
 
 
-class GetPoolInfoAvro(IMessageAvro):
+class PoolInfoAvroById(RequestByIdRecord):
     """
     Implements Avro representation of a Get Pool Info Message
     """
-    # Use __slots__ to explicitly declare all data members.
-    __slots__ = ["name", "message_id", "guid", "auth", "broker_id", "callback_topic", "id_token", "id"]
-
     def __init__(self):
+        super().__init__()
         self.name = IMessageAvro.GetPoolInfoRequest
-        self.message_id = None
-        self.guid = None
-        self.auth = None
-        self.callback_topic = None
-        self.broker_id = None
-        self.id_token = None
-        # Unique id used to track produce request success/failures.
-        # Do *not* include in the serialized object.
-        self.id = uuid4()
-
-    def from_dict(self, value: dict):
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide conversion from dict to our class for de-serialization
-        :param value: incoming message dictionary
-        """
-        if value['name'] != IMessageAvro.GetPoolInfoRequest:
-            raise Exception("Invalid message")
-        self.message_id = value.get('message_id')
-        self.guid = value.get('guid')
-        self.callback_topic = value.get('callback_topic')
-        self.broker_id = value.get('broker_id', None)
-        self.id_token = value.get('id_token', None)
-
-        if value.get('auth', None) is not None:
-            self.auth = AuthAvro()
-            self.auth.from_dict(value['auth'])
-
-    def to_dict(self) -> dict:
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide a dict representation of our class for serialization.
-        :return dict representing the class
-        """
-        if not self.validate():
-            raise Exception("Invalid arguments")
-        result = {
-            "name": self.name,
-            "message_id": self.message_id,
-            "guid": self.guid,
-            "callback_topic": self.callback_topic,
-            "broker_id": self.broker_id,
-            "id_token": self.id_token
-        }
-        if self.auth is not None:
-            result['auth'] = self.auth.to_dict()
-
-        return result
-
-    def get_message_id(self) -> str:
-        """
-        Returns the message_id
-        """
-        return self.message_id
-
-    def get_message_name(self) -> str:
-        return self.name
-
-    def get_callback_topic(self) -> str:
-        return self.callback_topic
-
-    def get_broker_id(self) -> str:
-        """
-        Returns broker id
-        @param broker id
-        """
-        return self.broker_id
-
-    def get_id(self) -> str:
-        return self.id.__str__()
-
-    def get_id_token(self) -> str:
-        """
-        Returns Identity Token used for PDP authorization
-        @param identity token
-        """
-        return self.id_token
-
-    def __str__(self):
-        return "name: {} message_id: {} guid: {} auth: {} callback_topic: {} broker_id: {} id_token: {}".format(
-            self.name, self.message_id, self.guid, self.auth, self.callback_topic, self.broker_id, self.id_token)
