+++ fabric/message_bus/messages/get_unit_avro.py
@@ -26,111 +26,17 @@
 """
 Implements Avro representation of a Get Unit Message
 """
-from uuid import uuid4
-
-from fabric.message_bus.messages.auth_avro import AuthAvro
+from fabric.message_bus.messages.request_by_id_record import RequestByIdRecord
 from fabric.message_bus.messages.message import IMessageAvro
 
 
-class GetUnitAvro(IMessageAvro):
+class UnitAvroById(RequestByIdRecord):
     """
     Implements Avro representation of a Get Unit Message
     """
-    # Use __slots__ to explicitly declare all data members.
-    __slots__ = ["name", "message_id", "guid", "auth", "unit_id", "callback_topic", "id_token", "id"]
-
     def __init__(self):
+        super().__init__()
         self.name = IMessageAvro.GetUnitRequest
-        self.message_id = None
-        self.guid = None
-        self.auth = None
-        self.unit_id = None
-        self.callback_topic = None
-        self.id_token = None
-        # Unique id used to track produce request success/failures.
-        # Do *not* include in the serialized object.
-        self.id = uuid4()
-
-    def from_dict(self, value: dict):
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide conversion from dict to our class for de-serialization
-        :param value: incoming message dictionary
-        """
-        if value['name'] != IMessageAvro.GetUnitRequest:
-            raise Exception("Invalid message")
-        self.message_id = value['message_id']
-        self.guid = value['guid']
-        self.callback_topic = value['callback_topic']
-        self.unit_id = value.get("unit_id", None)
-        self.id_token = value.get("id_token", None)
-
-        if value.get('auth', None) is not None:
-            self.auth = AuthAvro()
-            self.auth.from_dict(value['auth'])
-
-    def to_dict(self) -> dict:
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide a dict representation of our class for serialization.
-        :return dict representing the class
-        """
-        if not self.validate():
-            raise Exception("Invalid arguments")
-
-        result = {
-            "name": self.name,
-            "message_id": self.message_id,
-            "guid": self.guid,
-            "callback_topic": self.callback_topic,
-            "id_token": self.id_token
-        }
-        if self.auth is not None:
-            result['auth'] = self.auth.to_dict()
-
-        if self.unit_id is not None:
-            result['unit_id'] = self.unit_id
-        return result
-
-    def get_message_id(self) -> str:
-        """
-        Returns the message_id
-        """
-        return self.message_id
-
-    def get_message_name(self) -> str:
-        """
-        Returns the message name
-        """
-        return self.name
-
-    def get_callback_topic(self) -> str:
-        """
-        Returns the callback topic
-        """
-        return self.callback_topic
-
-    def get_id(self) -> str:
-        """
-        Returns the id
-        """
-        return self.id.__str__()
-
-    def get_unit_id(self) -> str:
-        """
-        Returns the unit id
-        """
-        return self.unit_id
-
-    def get_id_token(self) -> str:
-        """
-        Returns the id token
-        """
-        return self.id_token
-
-    def __str__(self):
-        return "name: {} message_id: {} guid: {} auth: {} unit_id: {} callback_topic: {} id_token: {}".format(
-            self.name, self.message_id, self.guid, self.auth, self.unit_id, self.callback_topic, self.id_token)
 
     def validate(self) -> bool:
         """
