+++ fabric/message_bus/messages/result_delegation_avro.py
@@ -26,102 +26,14 @@
 """
 Implements Avro representation of a Result Message containing Delegations
 """
-from uuid import uuid4
-
-
-from fabric.message_bus.messages.delegation_avro import DelegationAvro
-from fabric.message_bus.messages.result_avro import ResultAvro
 from fabric.message_bus.messages.message import IMessageAvro
+from fabric.message_bus.messages.result_record_list import ResultRecordList
 
 
-class ResultDelegationAvro(IMessageAvro):
+class ResultDelegationAvro(ResultRecordList):
     """
     Implements Avro representation of a Result Message containing Delegations
     """
-    # Use __slots__ to explicitly declare all data members.
-    __slots__ = ["name", "message_id", "status", "delegations", "id"]
-
     def __init__(self):
+        super().__init__()
         self.name = IMessageAvro.ResultDelegation
-        self.message_id = None
-        self.status = None
-        self.delegations = None
-        # Unique id used to track produce request success/failures.
-        # Do *not* include in the serialized object.
-        self.id = uuid4()
-
-    def from_dict(self, value: dict):
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide conversion from dict to our class for de-serialization
-        :param value: incoming message dictionary
-        """
-        if value['name'] != IMessageAvro.ResultDelegation:
-            raise Exception("Invalid message")
-        self.message_id = value['message_id']
-        self.status = ResultAvro()
-        self.status.from_dict(value['status'])
-        delegations_list = value.get('delegations', None)
-        if delegations_list is not None:
-            for d in delegations_list:
-                del_obj = DelegationAvro()
-                del_obj.from_dict(d)
-                if self.delegations is None:
-                    self.delegations = []
-                self.delegations.append(del_obj)
-
-    def to_dict(self) -> dict:
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide a dict representation of our class for serialization.
-        :return dict representing the class
-        """
-        if not self.validate():
-            raise Exception("Invalid arguments")
-
-        result = {
-            "name": self.name,
-            "message_id": self.message_id,
-            "status": self.status.to_dict()
-        }
-        if self.delegations is not None:
-            temp = []
-            for s in self.delegations:
-                temp.append(s.to_dict())
-            result["delegations"] = temp
-        return result
-
-    def get_message_id(self) -> str:
-        """
-        Returns the message_id
-        """
-        return self.message_id
-
-    def get_message_name(self) -> str:
-        return self.name
-
-    def __str__(self):
-        return "name: {} message_id: {} status: {} delegations: {}".format(self.name, self.message_id, self.status,
-                                                                           self.delegations)
-
-    def get_id(self) -> str:
-        return self.id.__str__()
-
-    def set_status(self, status: ResultAvro):
-        """
-        Set status
-        """
-        self.status = status
-
-    def get_callback_topic(self) -> str:
-        return None
-
-    def validate(self) -> bool:
-        """
-        Check if the object is valid and contains all mandatory fields
-        :return True on success; False on failure
-        """
-        ret_val = super().validate()
-        if self.status is None:
-            ret_val = False
-        return ret_val
