+++ fabric/message_bus/messages/result_proxy_avro.py
@@ -26,116 +26,14 @@
 """
 Implements Avro representation of a Result Message containing Proxies
 """
-from typing import List
-from uuid import uuid4
-
-from fabric.message_bus.messages.proxy_avro import ProxyAvro
-from fabric.message_bus.messages.result_avro import ResultAvro
 from fabric.message_bus.messages.message import IMessageAvro
+from fabric.message_bus.messages.result_record_list import ResultRecordList
 
 
-class ResultProxyAvro(IMessageAvro):
+class ResultProxyAvro(ResultRecordList):
     """
     Implements Avro representation of a Result Message containing Proxies
     """
-    # Use __slots__ to explicitly declare all data members.
-    __slots__ = ["name", "message_id", "status", "proxies", "id"]
-
     def __init__(self):
+        super().__init__()
         self.name = IMessageAvro.ResultProxy
-        self.message_id = None
-        self.status = None
-        self.proxies = None
-        # Unique id used to track produce request success/failures.
-        # Do *not* include in the serialized object.
-        self.id = uuid4()
-
-    def from_dict(self, value: dict):
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide conversion from dict to our class for de-serialization
-        :param value: incoming message dictionary
-        """
-        if value['name'] != IMessageAvro.ResultProxy:
-            raise Exception("Invalid message")
-        self.message_id = value['message_id']
-        self.status = ResultAvro()
-        self.status.from_dict(value['status'])
-        proxies_list = value.get('proxies', None)
-        if proxies_list is not None:
-            for p in proxies_list:
-                proxy_obj = ProxyAvro()
-                proxy_obj.from_dict(p)
-                if self.proxies is None:
-                    self.proxies = []
-                self.proxies.append(proxy_obj)
-
-    def to_dict(self) -> dict:
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide a dict representation of our class for serialization.
-        :return dict representing the class
-        """
-        if not self.validate():
-            raise Exception("Invalid arguments")
-
-        result = {
-            "name": self.name,
-            "message_id": self.message_id,
-            "status": self.status.to_dict()
-        }
-        if self.proxies is not None:
-            temp = []
-            for s in self.proxies:
-                temp.append(s.to_dict())
-            result["proxies"] = temp
-        return result
-
-    def get_message_id(self) -> str:
-        """
-        Returns the message_id
-        """
-        return self.message_id
-
-    def get_message_name(self) -> str:
-        return self.name
-
-    def __str__(self):
-        return "name: {} message_id: {} status: {} proxies: {}".format(self.name, self.message_id, self.status,
-                                                                       self.proxies)
-
-    def get_status(self) -> ResultAvro:
-        """
-        Return status
-        @return status
-        """
-        return self.status
-
-    def set_status(self, value: ResultAvro):
-        """
-        Set status
-        @param value value
-        """
-        self.status = value
-
-    def get_proxies(self) -> List[ProxyAvro]:
-        """
-        Return proxies
-        """
-        return self.proxies
-
-    def get_id(self) -> str:
-        return self.id.__str__()
-
-    def get_callback_topic(self) -> str:
-        return None
-
-    def validate(self) -> bool:
-        """
-        Check if the object is valid and contains all mandatory fields
-        :return True on success; False on failure
-        """
-        ret_val = super().validate()
-        if self.status is None:
-            ret_val = False
-        return ret_val
