+++ fabric/message_bus/messages/result_slice_avro.py
@@ -26,117 +26,14 @@
 """
 Implements Avro representation of a Result Message containing Slices
 """
-from typing import List
-from uuid import uuid4
-
-from fabric.message_bus.messages.result_avro import ResultAvro
-from fabric.message_bus.messages.slice_avro import SliceAvro
+from fabric.message_bus.messages.result_record_list import ResultRecordList
 from fabric.message_bus.messages.message import IMessageAvro
 
 
-class ResultSliceAvro(IMessageAvro):
+class ResultSliceAvro(ResultRecordList):
     """
     Implements Avro representation of a Result Message containing Slices
     """
-    # Use __slots__ to explicitly declare all data members.
-    __slots__ = ["name", "message_id", "status", "slices", "id"]
-
     def __init__(self):
+        super().__init__()
         self.name = IMessageAvro.ResultSlice
-        self.message_id = None
-        self.status = None
-        self.slices = None
-        # Unique id used to track produce request success/failures.
-        # Do *not* include in the serialized object.
-        self.id = uuid4()
-
-    def from_dict(self, value: dict):
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide conversion from dict to our class for de-serialization
-        :param value: incoming message dictionary
-        """
-        if value['name'] != IMessageAvro.ResultSlice:
-            raise Exception("Invalid message")
-        self.message_id = value['message_id']
-        self.status = ResultAvro()
-        self.status.from_dict(value['status'])
-        slices_list = value.get('slices', None)
-        if slices_list is not None:
-            for s in slices_list:
-                slice_obj = SliceAvro()
-                slice_obj.from_dict(s)
-                if self.slices is None:
-                    self.slices = []
-                self.slices.append(slice_obj)
-
-    def to_dict(self) -> dict:
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide a dict representation of our class for serialization.
-        :return dict representing the class
-        """
-        if not self.validate():
-            raise Exception("Invalid arguments")
-
-        result = {
-            "name": self.name,
-            "message_id": self.message_id,
-            "status": self.status.to_dict()
-        }
-        if self.slices is not None:
-            temp = []
-            for s in self.slices:
-                temp.append(s.to_dict())
-            result["slices"] = temp
-        return result
-
-    def get_message_id(self) -> str:
-        """
-        Returns the message_id
-        """
-        return self.message_id
-
-    def get_message_name(self) -> str:
-        return self.name
-
-    def __str__(self):
-        return "name: {} message_id: {} status: {} slices: {}".format(self.name, self.message_id, self.status,
-                                                                      self.slices)
-
-    def get_status(self) -> ResultAvro:
-        """
-        Returns response status
-        @return response status
-        """
-        return self.status
-
-    def set_status(self, value: ResultAvro):
-        """
-        Set response status
-        @param value value
-        """
-        self.status = value
-
-    def get_slices(self) -> List[SliceAvro]:
-        """
-        Returns list of Slices
-        @param list of Slices
-        """
-        return self.slices
-
-    def get_id(self) -> str:
-        return self.id.__str__()
-
-    def get_callback_topic(self) -> str:
-        return None
-
-    def validate(self) -> bool:
-        """
-        Check if the object is valid and contains all mandatory fields
-        :return True on success; False on failure
-        """
-        ret_val = super().validate()
-        if self.status is None:
-            ret_val = False
-        return ret_val
