+++ fabric/message_bus/messages/result_unit_avro.py
@@ -26,114 +26,14 @@
 """
 Implements Avro representation of a Result Message containing Units
 """
-from typing import List
-from uuid import uuid4
-
 from fabric.message_bus.messages.message import IMessageAvro
-from fabric.message_bus.messages.result_avro import ResultAvro
-from fabric.message_bus.messages.unit_avro import UnitAvro
+from fabric.message_bus.messages.result_record_list import ResultRecordList
 
 
-class ResultUnitAvro(IMessageAvro):
+class ResultUnitAvro(ResultRecordList):
     """
     Implements Avro representation of a Result Message containing Units
     """
-    # Use __slots__ to explicitly declare all data members.
-    __slots__ = ["name", "message_id", "units", "callback_topic", "id"]
-
     def __init__(self):
+        super().__init__()
         self.name = IMessageAvro.ResultUnits
-        self.message_id = None
-        self.status = None
-        self.units = None
-        # Unique id used to track produce request success/failures.
-        # Do *not* include in the serialized object.
-        self.id = uuid4()
-
-    def get_status(self) -> ResultAvro:
-        """
-        Return result status
-        """
-        return self.status
-
-    def set_status(self, value: ResultAvro):
-        """
-        Set status
-        @param value value
-        """
-        self.status = value
-
-    def get_units(self) -> List[UnitAvro]:
-        """
-        Return List of units
-        """
-        return self.units
-
-    def from_dict(self, value: dict):
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide conversion from dict to our class for de-serialization
-        :param value: incoming message dictionary
-        """
-        if value['name'] != IMessageAvro.ResultUnits:
-            raise Exception("Invalid message")
-        self.message_id = value['message_id']
-        self.status = ResultAvro()
-        self.status.from_dict(value['status'])
-        temp_units = value.get('units', None)
-        if temp_units is not None:
-            self.units = []
-            for u in temp_units:
-                unit = UnitAvro()
-                unit.from_dict(u)
-                self.units.append(unit)
-
-    def to_dict(self) -> dict:
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide a dict representation of our class for serialization.
-        :return dict representing the class
-        """
-        if not self.validate():
-            raise Exception("Invalid arguments")
-
-        result = {
-            "name": self.name,
-            "message_id": self.message_id,
-            "status": self.status.to_dict()
-        }
-        if self.units is not None:
-            temp = []
-            for u in self.units:
-                temp.append(u.to_dict())
-            result["units"] = temp
-        return result
-
-    def get_message_id(self) -> str:
-        """
-        Returns the message_id
-        """
-        return self.message_id
-
-    def get_message_name(self) -> str:
-        return self.name
-
-    def get_id(self) -> str:
-        return self.id.__str__()
-
-    def __str__(self):
-        return "name: {} message_id: {} status: {} units: {}".format(
-            self.name, self.message_id, self.status, self.units)
-
-    def get_callback_topic(self) -> str:
-        return None
-
-    def validate(self) -> bool:
-        """
-        Check if the object is valid and contains all mandatory fields
-        :return True on success; False on failure
-        """
-        ret_val = super().validate()
-        if self.status is None:
-            ret_val = False
-        return ret_val
