+++ fabric/message_bus/messages/update_lease_avro.py
@@ -26,11 +26,11 @@
 """
 Implements Avro representation of a Update Lease Message
 """
-from fabric.message_bus.messages.message_base import MessageBaseAvro
+from fabric.message_bus.messages.reservation_or_delegation_record import ReservationOrDelegationRecord
 from fabric.message_bus.messages.message import IMessageAvro
 
 
-class UpdateLeaseAvro(MessageBaseAvro):
+class UpdateLeaseWithReservationOrDelegationRecord(ReservationOrDelegationRecord):
     """
     Implements Avro representation of a Update Lease Message
     """
