+++ fabric/message_bus/messages/update_reservation_avro.py
@@ -26,115 +26,16 @@
 """
 Implements Avro representation of a Update Reservation Message
 """
-from uuid import uuid4
 
-from fabric.message_bus.messages.auth_avro import AuthAvro
-from fabric.message_bus.messages.reservation_mng import ReservationMng
+from fabric.message_bus.messages.add_update_reservation_record import AddUpdateReservationRecord
 from fabric.message_bus.messages.message import IMessageAvro
 
 
-class UpdateReservationAvro(IMessageAvro):
+class UpdateReservationAvro(AddUpdateReservationRecord):
     """
     Implements Avro representation of a Update Reservation Message
     """
-    # Use __slots__ to explicitly declare all data members.
-    __slots__ = ["name", "message_id", "guid", "auth", "reservation_obj", "callback_topic", "id"]
 
     def __init__(self):
+        super().__init__()
         self.name = IMessageAvro.UpdateReservation
-        self.message_id = None
-        self.guid = None
-        self.auth = None
-        self.reservation_obj = None
-        self.callback_topic = None
-        # Unique id used to track produce request success/failures.
-        # Do *not* include in the serialized object.
-        self.id = uuid4()
-
-    def from_dict(self, value: dict):
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide conversion from dict to our class for de-serialization
-        :param value: incoming message dictionary
-        """
-        if value['name'] != IMessageAvro.UpdateReservation:
-            raise Exception("Invalid message")
-        self.message_id = value['message_id']
-        self.guid = value['guid']
-        self.callback_topic = value['callback_topic']
-
-        if value.get("reservation_obj", None) is not None:
-            res_value = value.get("reservation_obj", None)
-            self.reservation_obj = ReservationMng()
-            self.reservation_obj.from_dict(res_value)
-
-        if value.get('auth', None) is not None:
-            self.auth = AuthAvro()
-            self.auth.from_dict(value['auth'])
-
-    def to_dict(self) -> dict:
-        """
-        The Avro Python library does not support code generation.
-        For this reason we must provide a dict representation of our class for serialization.
-        :return dict representing the class
-        """
-        if not self.validate():
-            raise Exception("Invalid arguments")
-
-        result = {
-            "name": self.name,
-            "message_id": self.message_id,
-            "guid": self.guid,
-            "callback_topic": self.callback_topic
-        }
-        if self.auth is not None:
-            result['auth'] = self.auth.to_dict()
-
-        if self.reservation_obj is not None:
-            result['reservation_obj'] = self.reservation_obj.to_dict()
-
-        return result
-
-    def get_message_id(self) -> str:
-        """
-        Returns the message_id
-        """
-        return self.message_id
-
-    def get_message_name(self) -> str:
-        """
-        Returns the message name
-        """
-        return self.name
-
-    def get_callback_topic(self) -> str:
-        """
-        Returns the callback topic
-        """
-        return self.callback_topic
-
-    def get_id(self) -> str:
-        """
-        Returns the id
-        """
-        return self.id.__str__()
-
-    def get_reservation(self) -> ReservationMng:
-        """
-        Returns the reservation object
-        """
-        return self.reservation_obj
-
-    def __str__(self):
-        return "name: {} message_id: {} guid: {} auth: {} reservation_obj: {} callback_topic: {}".format(
-            self.name, self.message_id, self.guid, self.auth, self.reservation_obj, self.callback_topic)
-
-    def validate(self) -> bool:
-        """
-        Check if the object is valid and contains all mandatory fields
-        :return True on success; False on failure
-        """
-        ret_val = super().validate()
-        if self.auth is None or self.callback_topic is None or self.guid is None or self.reservation_obj is None:
-            ret_val = False
-        return ret_val
